Перенёс `BaseUpdate` и `MaxUpdate` из `maxo/types/base_update.py` в `maxo/types/base.py`
