+++ src/maxo/dialogs/context/intent_filter.py
@@ -3,7 +3,7 @@ from maxo.dialogs.api.internal import CONTEXT_KEY
 from maxo.fsm import StatesGroup
 from maxo.routing.ctx import Ctx
 from maxo.routing.filters import BaseFilter
-from maxo.types.base_update import BaseUpdate
+from maxo.types.base import BaseUpdate
 
 
 class IntentFilter(BaseFilter[BaseUpdate]):
