+++ src/maxo/routing/dispatcher.py
@@ -18,7 +18,7 @@ from maxo.routing.signals.base import BaseSignal
 from maxo.routing.signals.update import MaxoUpdate
 from maxo.routing.utils._resolving_inner_middlewares import resolve_middlewares
 from maxo.routing.utils.validate_router_graph import validate_router_graph
-from maxo.types.base_update import BaseUpdate
+from maxo.types.base import BaseUpdate
 
 
 class Dispatcher(Router):
