+++ src/maxo/routing/filters/sync.py
@@ -4,7 +4,7 @@ from typing import TypeVar
 
 from maxo.routing.ctx import Ctx
 from maxo.routing.filters.base import BaseFilter
-from maxo.types.base_update import BaseUpdate
+from maxo.types.base import BaseUpdate
 
 _UpdateT = TypeVar("_UpdateT", bound=BaseUpdate)
