+++ src/maxo/routing/handlers/update.py
@@ -7,7 +7,7 @@ from maxo.routing.ctx import Ctx
 from maxo.routing.filters.logic import combine_filters
 from maxo.routing.interfaces.filter import Filter
 from maxo.routing.interfaces.handler import Handler
-from maxo.types.base_update import BaseUpdate
+from maxo.types.base import BaseUpdate
 
 _UpdateT = TypeVar("_UpdateT", bound=BaseUpdate)
 _ReturnT_co = TypeVar("_ReturnT_co", covariant=True)
