+++ src/maxo/routing/interfaces/handler.py
@@ -2,7 +2,7 @@ from abc import abstractmethod
 from typing import Protocol, TypeVar
 
 from maxo.routing.ctx import Ctx
-from maxo.types.base_update import BaseUpdate
+from maxo.types.base import BaseUpdate
 
 _UpdateT = TypeVar("_UpdateT", bound=BaseUpdate)
 _ReturnT_co = TypeVar("_ReturnT_co", covariant=True)
