+++ src/maxo/routing/interfaces/middleware.py
@@ -2,7 +2,7 @@ from abc import abstractmethod
 from typing import Any, Protocol, TypeVar, runtime_checkable
 
 from maxo.routing.ctx import Ctx
-from maxo.types.base_update import BaseUpdate
+from maxo.types.base import BaseUpdate
 
 _UpdateT_co = TypeVar("_UpdateT_co", bound=BaseUpdate, covariant=True)
 _UpdateT_contra = TypeVar("_UpdateT_contra", bound=BaseUpdate, contravariant=True)
