+++ tests/maxo/routing/filters/test_logic.py
@@ -7,7 +7,7 @@ from maxo.routing.filters.logic import (
     combine_filters,
 )
 from maxo.routing.interfaces.filter import Filter
-from maxo.types.base_update import BaseUpdate
+from maxo.types.base import BaseUpdate
 
 TrueF = AlwaysTrueFilter
 FalseF = AlwaysFalseFilter
